fix: busy store usage
