+++ frontend/src/lib/components/accounts/HardwareWalletConnect.svelte
@@ -10,7 +10,7 @@
   } from "$lib/types/add-account.context";
   import { createEventDispatcher, getContext } from "svelte";
   import type { LedgerIdentity } from "$lib/identities/ledger.identity";
-  import { busy, startBusy, stopBusy } from "$lib/stores/busy.store";
+  import { busy, startBusy, stopBusy } from "@dfinity/gix-components";
 
   let connectionState: LedgerConnectionState =
     LedgerConnectionState.NOT_CONNECTED;
