+++ frontend/src/lib/components/accounts/HardwareWalletShowActionButton.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { i18n } from "$lib/stores/i18n";
   import { showAddressAndPubKeyOnHardwareWalletProxy } from "$lib/proxy/ledger.services.proxy";
-  import { busy } from "$lib/stores/busy.store";
+  import { busy } from "@dfinity/gix-components";
 
   const showInfo = async () =>
     await showAddressAndPubKeyOnHardwareWalletProxy();
