+++ frontend/src/lib/components/canister-detail/NewControllerReview.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { busy, startBusy, stopBusy } from "$lib/stores/busy.store";
+  import { busy, startBusy, stopBusy } from "@dfinity/gix-components";
   import { i18n } from "$lib/stores/i18n";
   import type { Principal } from "@dfinity/principal";
   import { createEventDispatcher, getContext } from "svelte";
