+++ frontend/src/lib/components/neurons/ConfirmDissolveDelay.svelte
@@ -11,10 +11,9 @@
     neuronStake,
     votingPower,
   } from "$lib/utils/neuron.utils";
-  import { busy, stopBusy } from "$lib/stores/busy.store";
   import { startBusyNeuron } from "$lib/services/busy.services";
   import { valueSpan } from "$lib/utils/utils";
-  import { Html } from "@dfinity/gix-components";
+  import { Html, busy, stopBusy } from "@dfinity/gix-components";
 
   export let delayInSeconds: number;
   export let neuron: NeuronInfo;
