+++ frontend/src/lib/components/neurons/ConfirmNeuronsMerge.svelte
@@ -4,14 +4,13 @@
   import { MAX_NEURONS_MERGED } from "$lib/constants/neurons.constants";
   import { startBusyNeuron } from "$lib/services/busy.services";
   import { mergeNeurons } from "$lib/services/neurons.services";
-  import { busy, stopBusy } from "$lib/stores/busy.store";
   import { i18n } from "$lib/stores/i18n";
   import { toastsError, toastsSuccess } from "$lib/stores/toasts.store";
   import { replacePlaceholders } from "$lib/utils/i18n.utils";
   import { formatToken } from "$lib/utils/token.utils";
   import { neuronStake } from "$lib/utils/neuron.utils";
   import { valueSpan } from "$lib/utils/utils";
-  import { Html } from "@dfinity/gix-components";
+  import { Html, busy, stopBusy } from "@dfinity/gix-components";
 
   export let neurons: NeuronInfo[];
