+++ frontend/src/lib/components/sns-neuron-detail/SnsNeuronHotkeysCard.svelte
@@ -11,7 +11,7 @@
   } from "@dfinity/gix-components";
   import { removeHotkey } from "$lib/services/sns-neurons.services";
   import { authStore } from "$lib/stores/auth.store";
-  import { startBusy, stopBusy } from "$lib/stores/busy.store";
+  import { startBusy, stopBusy } from "@dfinity/gix-components";
   import { i18n } from "$lib/stores/i18n";
   import { snsProjectSelectedStore } from "$lib/derived/selected-project.derived";
   import {
