+++ frontend/src/lib/modals/accounts/IcpTransactionModal.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import { createEventDispatcher } from "svelte";
   import { transferICP } from "$lib/services/accounts.services";
-  import { startBusy, stopBusy } from "$lib/stores/busy.store";
   import { i18n } from "$lib/stores/i18n";
   import { toastsSuccess } from "$lib/stores/toasts.store";
   import type { Account } from "$lib/types/account";
@@ -9,7 +8,11 @@
   import { isAccountHardwareWallet } from "$lib/utils/accounts.utils";
   import TransactionModal from "./NewTransaction/TransactionModal.svelte";
   import { OWN_CANISTER_ID } from "$lib/constants/canister-ids.constants";
-  import type { WizardStep } from "@dfinity/gix-components";
+  import {
+    stopBusy,
+    startBusy,
+    type WizardStep,
+  } from "@dfinity/gix-components";
   import { mainTransactionFeeStoreAsToken } from "$lib/derived/main-transaction-fee.derived";
 
   export let selectedAccount: Account | undefined = undefined;
