+++ frontend/src/lib/modals/sns/SwapModal/ParticipateSwapModal.svelte
@@ -15,7 +15,7 @@
   import type { SnsSummary, SnsSwapCommitment } from "$lib/types/sns";
   import TransactionModal from "$lib/modals/accounts/NewTransaction/TransactionModal.svelte";
   import { nonNullish } from "$lib/utils/utils";
-  import { startBusy, stopBusy } from "$lib/stores/busy.store";
+  import { startBusy, stopBusy } from "@dfinity/gix-components";
   import {
     getSwapAccount,
     participateInSwap,
