+++ frontend/src/lib/stores/debug.store.ts
@@ -4,10 +4,9 @@ import type { HardwareWalletNeuronsStore } from "$lib/types/hardware-wallet-neur
 import type { ProjectDetailStore } from "$lib/types/project-detail.context";
 import type { SelectedAccountStore } from "$lib/types/selected-account.context";
 import type { SelectedProposalStore } from "$lib/types/selected-proposal.context";
-import { toastsStore } from "@dfinity/gix-components";
+import { busyStore, toastsStore } from "@dfinity/gix-components";
 import { derived, readable, type Readable, type Writable } from "svelte/store";
 import { accountsStore } from "./accounts.store";
-import { busyStore } from "./busy.store";
 import { canistersStore } from "./canisters.store";
 import { knownNeuronsStore } from "./knownNeurons.store";
 import { neuronsStore } from "./neurons.store";
