+++ frontend/src/tests/lib/services/busy.services.spec.ts
@@ -3,8 +3,8 @@
  */
 import { startBusyNeuron } from "$lib/services/busy.services";
 import { accountsStore } from "$lib/stores/accounts.store";
-import * as busyStore from "$lib/stores/busy.store";
 import { neuronsStore } from "$lib/stores/neurons.store";
+import * as busyStore from "@dfinity/gix-components";
 import {
   mockHardwareWalletAccount,
   mockMainAccount,
